Update based on comments from @CGodiksen
