+++ server/src/main.rs
@@ -95,6 +95,9 @@ fn main() -> Result<(), String> {
         });
 
         // Register tables and model tables.
+        context.metadata_manager.register_tables(&context)
+            .map_err(|error| format!("Unable to register tables: {}", error))?;
+
         context.metadata_manager.register_model_tables(&context)
             .map_err(|error| format!("Unable to register model tables: {}", error))?;
