+++ server/src/types.rs
@@ -26,7 +26,6 @@ pub type ArrowTimeSeriesId = datafusion::arrow::datatypes::UInt64Type;
 
 // Types used for a collection of time series ids.
 pub type TimeSeriesIdBuilder = datafusion::arrow::array::PrimitiveBuilder<ArrowTimeSeriesId>;
-pub type TimeSeriesIdArray = datafusion::arrow::array::PrimitiveArray<ArrowTimeSeriesId>;
 
 // Types used for a single timestamp.
 pub type Timestamp = std::primitive::i64; // It is signed to match TimestampMillisecondType.
